+++ TrabalhoAlgebraLinear/src/model/Matriz.java
@@ -3,17 +3,17 @@ package model;
 public class Matriz {
 	private int linha;
 	private int coluna;
-	private float[][] matriz;
+	private double[][] matriz;
 	
 	//Construtor que recebe o numero de linhas e colunas
 	public Matriz(int linha, int coluna) {
 		this.linha = linha;
 		this.coluna = coluna;
-		this.matriz = new float[linha][coluna];
+		this.matriz = new double[linha][coluna];
 	}
 	
 	//Construtor que recebe um array
-	public Matriz(float[][] array) {
+	public Matriz(double[][] array) {
 		this.linha = array.length;
 		this.coluna = array[0].length;
 		this.matriz = array;
@@ -33,17 +33,17 @@ public class Matriz {
 	}
 	
 	//Retorna a matriz 
-	public float[][] getMatriz(){
+	public double[][] getMatriz(){
 		return matriz;
 	}
 	
 	//Retorna um determinado valor da matriz 
-	public float value(int linha, int coluna) {
+	public double value(int linha, int coluna) {
 		return (linha>=0 && coluna>=0) ? matriz[linha][coluna] : 0;
 	}
 	
 	//Muda um valor de uma determinada posição da matriz
-	public void setValue(int linha, int coluna, float valor) {
+	public void setValue(int linha, int coluna, double valor) {
 		if(linha>=0 && coluna>=0)
 			matriz[linha][coluna] = valor;
 	}
