+++ TrabalhoAlgebraLinear/src/model/Sistema.java
@@ -3,11 +3,11 @@ package model;
 public class Sistema {
 	private int numEq; //número de equações    => Linhas
 	private int numIncog; //número de incógnitas => Colunas
-	private float[][] coef; //Matriz dos Coeficientes das incógnitas
-	private float[] termo; //Matriz dos Termos Independentes
+	private double[][] coef; //Matriz dos Coeficientes das incógnitas
+	private double[] termo; //Matriz dos Termos Independentes
 	
 	//Construtor: recebe tudo
-	public Sistema(int numEq, int numIncog, float[][] coef, float[] termo) {
+	public Sistema(int numEq, int numIncog, double[][] coef, double[] termo) {
 		this.numEq = numEq;
 		this.numIncog = numIncog;
 		this.coef = coef;
@@ -15,7 +15,7 @@ public class Sistema {
 	}
 	
 	//Construtor: recebe a Matriz Ampliada, e o numero de linhas e colunas
-	public Sistema(float[][] matrizAmp, int numEq, int numIncog ) {
+	public Sistema(double[][] matrizAmp, int numEq, int numIncog ) {
 		this.numEq = numEq;
 		this.numIncog = numIncog;
 		
@@ -38,22 +38,22 @@ public class Sistema {
 	}
 	
 	//Retorna a matriz dos coeficientes
-	public float[][] getMatrizCoef() {
+	public double[][] getMatrizCoef() {
 		return coef;
 	}
 	
 	//Retorna um determinado coeficiente da matriz dos coeficientes
-	public float getCoef(int linha,int coluna) {
+	public double getCoef(int linha,int coluna) {
 		return (linha>=0 && coluna>=0) ? coef[linha][coluna] : 0;
 	}
 	
 	//Retorna a Matriz dos Termos Independentes
-	public float[] getMatrizTermo() {
+	public double[] getMatrizTermo() {
 		return termo;
 	}
 	
 	//Retorna um determinado termo independente
-	public float getTermo(int eq) { //eq = equação correspondente
+	public double getTermo(int eq) { //eq = equação correspondente
 		return (eq>=0) ? termo[eq] : 0;
 	}
 
@@ -63,8 +63,8 @@ public class Sistema {
 	}
 	
 	//Retorna a Matriz Ampliada
-	public float[][] getMatrizAmpliada() {
-		float[][] matrizAmp = new float[numEq][numIncog+1];
+	public double[][] getMatrizAmpliada() {
+		double[][] matrizAmp = new double[numEq][numIncog+1];
 		for(int i = 0; i < numEq; i++) 
 			for(int j = 0; j < numIncog; j++)
 				matrizAmp[i][j] = coef[i][j];
