+++ TrabalhoAlgebraLinear/src/model/Vetor.java
@@ -1,19 +1,19 @@
 package model;
 
 public class Vetor {
-	private float[] coordenadas;
+	private double[] coordenadas;
 	
 	public Vetor(int numCoordenadas) {
-		this.coordenadas = new float[numCoordenadas];
+		this.coordenadas = new double[numCoordenadas];
 	}
 	
-	public Vetor(float[] coordenadas) {
+	public Vetor(double[] coordenadas) {
 		this.coordenadas = coordenadas;
 	}
 	
 	//Retorna o vetor atual normalizado
 	public Vetor normaliza() {
-		float norma = modulo();
+		double norma = modulo();
 		if(norma == 1)
 			return new Vetor(this.coordenadas);
 		
@@ -24,16 +24,16 @@ public class Vetor {
 	}
 	
 	//Retorna o modulo do vetor
-	public float modulo() {
-		float norma = 0;
+	public double modulo() {
+		double norma = 0;
 		for(int posCoordenada = 0; posCoordenada < coordenadas.length; posCoordenada++) 
 			norma += coordenadas[posCoordenada] * coordenadas[posCoordenada]; 
-		return (float) Math.sqrt(norma);
+		return (double) Math.sqrt(norma);
 	}
 	
 	//Retorna um vetor multiplicado por um Escalar
-	public Vetor multiplicaEscalar(float scalar) {
-		float[] vetorEscalonado = new float[this.coordenadas.length];
+	public Vetor multiplicaEscalar(double scalar) {
+		double[] vetorEscalonado = new double[this.coordenadas.length];
 		for(int posCoordenada = 0; posCoordenada < coordenadas.length; posCoordenada++) 
 			vetorEscalonado[posCoordenada] = scalar * coordenadas[posCoordenada]; 
 		
@@ -46,17 +46,17 @@ public class Vetor {
 	}
 	
 	//Retorna o valor de uma determinada coordenada do vetor
-	public float getValorCoordenada(int posCoordenada) {
+	public double getValorCoordenada(int posCoordenada) {
 		return (posCoordenada >= 0 && posCoordenada < coordenadas.length) ?
 				coordenadas[posCoordenada] : 0;
 	}
 	
-	//retorna todas as coordenadas como um array de float
-	public float[] getCoordenadas() {
+	//retorna todas as coordenadas como um array de double
+	public double[] getCoordenadas() {
 		return coordenadas;
 	}
 	
-	public void setCoordenadas(float[] coordenadas) {
+	public void setCoordenadas(double[] coordenadas) {
 		this.coordenadas = coordenadas;
 	}
